+++ lib/adhearsion/reporter.rb
@@ -19,6 +19,7 @@ module Adhearsion
       config :reporter do
         api_key nil,                  desc: "The Airbrake/Errbit API key"
         url     "http://airbrake.io", desc: "Base URL for notification service"
+        app_name "Reporter", desc: "Application name, used for reporting"
         notifier Adhearsion::Reporter::AirbrakeNotifier,
           desc: "The class that will act as the notifier. Built-in classes are Adhearsion::Reporter::AirbrakeNotifier and Adhearsion::Reporter::NewrelicNotifier",
           transform: Proc.new { |v| const_get(v.to_s) }
@@ -31,6 +32,7 @@ module Adhearsion
           developer_mode false, desc: "More information but very high overhead in memory"
           log_level 'info', desc: "The newrelic's agent log level"
         }
+        email Hash.new(via: :sendmail), desc: "Used to configure the email notifier"
       end
 
       init :reporter do