+++ lib/adhearsion/reporter.rb
@@ -6,6 +6,7 @@ require 'adhearsion'
 require 'adhearsion/reporter/airbrake_notifier'
 require 'adhearsion/reporter/newrelic_notifier'
 require 'adhearsion/reporter/email_notifier'
+require 'adhearsion/reporter/sentry_notifier'
 
 module Adhearsion
   class Reporter
@@ -21,7 +22,7 @@ module Adhearsion
         url     "http://airbrake.io", desc: "Base URL for notification service"
         app_name "Adhearsion", desc: "Application name, used for reporting"
         notifier Adhearsion::Reporter::AirbrakeNotifier,
-          desc: "The class that will act as the notifier. Built-in classes are Adhearsion::Reporter::AirbrakeNotifier and Adhearsion::Reporter::NewrelicNotifier",
+          desc: "The class that will act as the notifier. Built-in classes are Adhearsion::Reporter::AirbrakeNotifier, Adhearsion::Reporter::NewrelicNotifier, and Adhearsion::Reporter::SentryNotifier",
           transform: Proc.new { |v| const_get(v.to_s) }
         notifiers [],
           desc: "Collection of classes that will act as notifiers",
@@ -36,10 +37,16 @@ module Adhearsion
           log_level 'info', desc: "The newrelic's agent log level"
         }
         email Hash.new(via: :sendmail), desc: "Used to configure the email notifier, with options accepted by the pony (https://github.com/benprew/pony) gem"
+        sentry {
+          dsn 'https://<user>:<password>@app.getsentry.com/<application>', desc: "The SENTRY_DSN, or client key that has been created in Sentry"
+          current_environment 'production', 'The current execution environment'
+          environments ['production'], 'The environments for which Sentry is active'
+        }
+
       end
 
       init :reporter do
-        # If the notifiers is empty (the default), then use whatever content 
+        # If the notifiers is empty (the default), then use whatever content
         # was in the "notifier" property. This will allow the default, or an
         # explicitly set value to be used. Not that notifiers has been set, then
         # the notifier will be ignored.
@@ -48,7 +55,7 @@ module Adhearsion
           Events.register_callback(:exception) do |e, logger|
             Reporter.config.notifier.instance.notify e
           end
-        else 
+        else
           Reporter.config.notifiers.each do |notifier|
             notifier.init
             Events.register_callback(:exception) do |e, logger|