+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # develop
+  * Allow reporting of custom environment to Airbrake
+
+# v2.2.0
   * New notifiers: EmailNotifier and SentryNotifier
   * Allow configuring multiple notifiers as a list. The `notifier` option is deprecated in favour of `notifiers`, which can be used the same for a single notifier.